### 2017-11-21 #468 @arfon
@jronayne - please take another look at the [submission guidelines](http://joss.theoj.org/about#paper_structure) and make sure you have a `paper.md` and `paper.bib` file in your repository.
