### 2017-11-22 8dd0547
Delete JOSS_summary.tex
