+++ paper/JOSS_summary.tex
@@ -1,89 +1,0 @@
-\documentclass[10pt,
-%showpacs,
-amsmath,amssymb,
-aps,prd,nofootinbib,eqsecnum,a4paper]{revtex4}
-\usepackage{graphicx,epsf}
-\usepackage{float}
-\usepackage{wrapfig}
-%
-\usepackage{xcolor}
-\usepackage[colorlinks = true,
-            linkcolor = blue,
-            urlcolor  = blue,
-            citecolor = blue,
-            anchorcolor = blue]{hyperref}
-%
-\topmargin-10mm
-%
-%
-\def\be{\begin{equation}}
-\def\ee{\end{equation}}
-%
-\def\bea{\begin{eqnarray}}
-\def\eea{\end{eqnarray}}
-%
-\def\p{\partial}
-\def\n{\nabla}
-\def\dkk{\p_k\p^k}
-\def\sh{{\sigma}}
-\def\L{{\pounds}}
-
-\usepackage{framed}
-
-
-\usepackage{float}
-
-\newcommand{\CC}{C\nolinebreak\hspace{-.05em}\raisebox{.4ex}{\tiny\bf +}\nolinebreak\hspace{-.10em}\raisebox{.4ex}{\tiny\bf +}}
-\def\CC{{C\nolinebreak[4]\hspace{-.05em}\raisebox{.4ex}{\tiny\bf ++}}}
-\def\S{ }
-\begin{document}
-
-
-\title{PyTransport: A Python package for the calculation of inflationary correlation functions}
-\author{David J. Mulryne and John W. Ronayne}
-\affiliation{Astronomy Unit, School of Physics and Astronomy,
-Queen Mary University of London, Mile End Road, London, E1 4NS, UK}
-\email[]{d.mulryne@qmul.ac.uk , j.ronayne@qmul.ac.uk}
-
-\date{\today}
-
-
-\maketitle
-\begin{figure}[H]
-\centering
-\includegraphics[width=11cm]{PyTransLogo}
-\end{figure}
-\section{Summary}
-PyTransport constitutes a straightforward code written in \CC \S  together 
-with Python scripts which automatically edit, compile and run the \CC \S code as a 
-Python module. It has been written for Unix-like systems (OS X and Linux).
-The transport method we utilise means 
-only coupled differential equations need to be solved, and the implementation presented here 
-combines the speed of \CC \S  with the functionality and convenience of Python. 
-
-The code is intended to be a reusable resource for inflationary cosmology. It enables users to quickly create a 
-complied Python module(s) for any given model(s) of multi-field inflation. 
-Primarily the module employs the transport approach to inflationary cosmology to calculate 
-the tree-level power-spectrum and bispectrum of user specified models of multi-field inflation, 
-accounting for all sub and super-horizon effects. To this end,
-the module contains a number 
-functions that can be called from Python and that perform tasks such as calculating the background evolution 
-of the cosmology, as well as the evolution of the two and three point functions. We also provide a number of further functions written in 
-Python that perform common tasks such as calculating the power spectrum or bispectrum over a range of scales by utilising the 
-compiled module.
-The true power of the approach, however, is that users can rapidly write their own scripts, or adapt ours, to suit their own needs. 
-
-The transport approach to inflationary perturbation theory that the code employs 
-can be seen as the differential version of the integral expressions of the In-In formalism. It 
-is helpful numerically because it provides a set of ordinary differential equations for the correlation functions  
-of inflationary perturbations. The code solves these equations from deep inside the horizon until some desired time 
-after horizon crossing using a standard variable step size ordinary differential equation (ODE) 
-routine with error control. Such off the shelf 
-routines are extremely well tested, and provide
[…]
