### 2017-11-22 eefd3cf
Delete paper2.bib
