+++ paper/paper2.bib
@@ -1,344 +1,0 @@
-@article{xxx, 
- author     =  "Dias, Mafalda and Frazer, Jonathan
-                        and Mulryne, David and Seery, David",
-title          =  "Numerical evaluation of the bispectrum in multiple field inflation
-- the transport approach with code",
-year = "2016"
-}
-@article{xxx1, 
- author     =  "Ronayne, John 
-                        and Mulryne, David",
-title          =  "Numerically evaluating the bispectrum in curved field-space
- -- with PyTransport 2.0",
-year = "2017"
-}
-@article{xxx2, 
-author = "Seery, David",
-title = "CppTransport: a platform to automate calculations of inflationary correlation functions",
-year = "2016",
-doi ="10.5281/zenodo.61239"
-}
-@article{Maldacena:2002vr,
-      author         = "Maldacena, Juan Martin",
-      title          = "{Non-Gaussian features of primordial fluctuations in
-                        single field inflationary models}",
-      journal        = "JHEP",
-      volume         = "0305",
-      pages          = "013",
-      year           = "2003",
-      eprint         = "astro-ph/0210603",
-      archivePrefix  = "arXiv",
-      primaryClass   = "astro-ph",
-}
-@article{Seery:2005wm,
-      author         = "Seery, David and Lidsey, James E.",
-      title          = "{Primordial non-Gaussianities in single field inflation}",
-      journal        = "JCAP",
-      volume         = "0506",
-      pages          = "003",
-      doi            = "10.1088/1475-7516/2005/06/003",
-      year           = "2005",
-      eprint         = "astro-ph/0503692",
-      archivePrefix  = "arXiv",
-      primaryClass   = "astro-ph",
-      SLACcitation   = "%%CITATION = ASTRO-PH/0503692;%%",
-}
-@article{Chen:2006nt,
-      author         = "Chen, Xingang and Huang, Min-xin and Kachru, Shamit and
-                        Shiu, Gary",
-      title          = "{Observational signatures and non-Gaussianities of
-                        general single field inflation}",
-      journal        = "JCAP",
-      volume         = "0701",
-      pages          = "002",
-      doi            = "10.1088/1475-7516/2007/01/002",
-      year           = "2007",
-      eprint         = "hep-th/0605045",
-      archivePrefix  = "arXiv",
-      primaryClass   = "hep-th",
-      reportNumber   = "SLAC-PUB-11840, MAD-TH-06-3, UFIFT-HEP-06-9,
-                        SU-ITP-06-12, CU-TP-1147",
-      SLACcitation   = "%%CITATION = HEP-TH/0605045;%%",
-}
-@article{Elliston:2012ab,
-      author         = "Elliston, Joseph and Seery, David and Tavakol, Reza",
-      title          = "{The inflationary bispectrum with curved field-space}",
-      journal        = "JCAP",
-      volume         = "1211",
-      pages          = "060",
-      doi            = "10.1088/1475-7516/2012/11/060",
-      year           = "2012",
-      eprint         = "1208.6011",
-      archivePrefix  = "arXiv",
-      primaryClass   = "astro-ph.CO",
-      SLACcitation   = "%%CITATION = ARXIV:1208.6011;%%",
-}
-@article{Rigopoulos:2003ak,
-      author         = "Rigopoulos, G.I. and Shellard, E.P.S.",
-      title          = "{The separate universe approach and the evolution of
-                        nonlinear superhorizon cosmological perturbations}",
-      journal        = "Phys.Rev.",
-      volume         = "D68",
-      pages          = "123518",
-      doi            = "10.1103/PhysRevD.68.123518",
[…]
