+++ paper/paper.bib
@@ -1,18 +1,18 @@
-@article{xxx, 
+@article{Dias, 
  author     =  "Dias, Mafalda and Frazer, Jonathan
                         and Mulryne, David and Seery, David",
 title          =  "Numerical evaluation of the bispectrum in multiple field inflation
 - the transport approach with code",
 year = "2016"
 }
-@article{xxx1, 
+@article{Ronayne, 
  author     =  "Ronayne, John 
                         and Mulryne, David",
 title          =  "Numerically evaluating the bispectrum in curved field-space
  -- with PyTransport 2.0",
 year = "2017"
 }
-@article{xxx2, 
+@article{Seery, 
 author = "Seery, David",
 title = "CppTransport: a platform to automate calculations of inflationary correlation functions",
 year = "2016",
