### 2017-11-23 #468 @jronayne
Hello,

Apologises, I had missed that part of the guideline. I have added the paper.md and paper.bib files into the repository.

It should be good to go.

Regards,
John
