### 2017-12-09 #468 @arfon
@joezuntz - would you be willing to review this submission for JOSS? You can read more about what's involved here: http://joss.theoj.org/about#reviewer_guidelines
