### 2018-01-09 jronayne/PyTransport#3 @joezuntz
[JOSS] Installation failure
Hi.  This is an issue as part of my review for JOSS.

I cloned the repository master branch and followed the instructions in Docs/UserGuide2.pdf

I tried to run the attached file (which github made me add ".txt" to the end of to be able to attach it).  I got the following result:

```
Joes-MacBook-Pro:PyTransport jaz$ python test1.py
running install
error: must supply either home or prefix/exec-prefix -- not both
Traceback (most recent call last):
  File "test1.py.txt", line 11, in <module>
    PyTransSetup.compileName('joe')
  File "/Users/jaz/src/PyTransport/PyTransport/PyTransSetup.py", line 115, in compileName
    shutil.rmtree(location+"/build/")
  File "/usr/local/Cellar/python/2.7.13_1/Frameworks/Python.framework/Versions/2.7/lib/python2.7/shutil.py", line 239, in rmtree
    onerror(os.listdir, path, sys.exc_info())
  File "/usr/local/Cellar/python/2.7.13_1/Frameworks/Python.framework/Versions/2.7/lib/python2.7/shutil.py", line 237, in rmtree
    names = os.listdir(path)
OSError: [Errno 2] No such file or directory: '/Users/jaz/src/PyTransport/PyTransport/PyTrans//build/'
```

When I made that directory and continued I got the first part of the error again:

```
Joes-MacBook-Pro:PyTransport jaz$ mkdir /Users/jaz/src/PyTransport/PyTransport/PyTrans//build
Joes-MacBook-Pro:PyTransport jaz$ python test1.py.txt
running install
error: must supply either home or prefix/exec-prefix -- not both
```

I also noticed that doing this modified some of the version-controlled files, which is pretty surprising and non-standard behaviour.  You might want to remove code that will be overwritten from version control.

```
Joes-MacBook-Pro:PyTransport jaz$ git status
On branch master
Your branch is up-to-date with 'origin/master'.
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git checkout -- <file>..." to discard changes in working directory)

	modified:   PyTransport/CppTrans/potential.h
	modified:   PyTransport/PyTrans/PyTrans.cpp
	modified:   PyTransport/PyTrans/moduleSetup.py
	modified:   PyTransport/PyTransSetup.pyc

```

Attached file:
[test1.py.txt](https://github.com/jronayne/PyTransport/files/1614952/test1.py.txt)
