### 2018-01-09 #494 @joezuntz
I've hit an installation issue that I've listed here: https://github.com/jronayne/PyTransport/issues/3
so I can't yet comment on the functionality.  I'll look at those parts of the docs that I can at this point.
