### 2018-01-09 jronayne/PyTransport#4 @joezuntz
[JOSS]  Automated tests
This JOSS review guidelines want me to check that there are automated tests in the code, for example using the python package "nose". There don't seem to be any such tests, though the examples are good.
