### 2018-01-09 #494 @joezuntz
The code is also missing automated tests: https://github.com/jronayne/PyTransport/issues/4
