### 2018-01-09 jronayne/PyTransport#5 @joezuntz
[JOSS] Contributions statement and documentation
JOSS requires that you have a statement with the code detailing this:
```
Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
```

I couldn't see anything about contributions, though the other things are mentioned on the website.
Also you might want to link to transportmethod.com in the README.md.
