### 2018-01-09 #494 @joezuntz
I've asked for a contributions statement: https://github.com/jronayne/PyTransport/issues/5
