### 2018-01-09 #494 @joezuntz
@arfon Can you clarify about the DOIs requirement - the bibtex file lists them, but do you want them somewhere more obvious than that?  e.g. in the paper PDF itself?
