### 2018-01-09 #494 @arfon
> @arfon Can you clarify about the DOIs requirement - the bibtex file lists them, but do you want them somewhere more obvious than that? e.g. in the paper PDF itself?

Basically we just want it present in the bibtex if there is a DOI available. That way we can generate better Crossref metadata when we publish the JOSS paper.
