### 2018-01-09 jronayne/PyTransport#3 @jronayne
Thank you very much for this and the time you are taking to review the code, and apologies it has not worked for you out of the box. The reason there is an error with the build folder is that folder is created when the installation takes place but because the installation is failing for you no folder is created and hence the error. The installation error itself "error: must supply either home or prefix/exec-prefix -- not both” we have not encountered before. A similar error seems to be documented to occur using pip install with a target directory with hombrew’s python installation. Could we ask what python installation you are using? And perhaps you could try the recommended fix https://stackoverflow.com/questions/24257803/distutilsoptionerror-must-supply-either-home-or-prefix-exec-prefix-not-both answer 19 — by adding the setup.cfg file to the PyTrans/ folder . We appreciate it is not your job to fix bugs, but this would help us to address this issue — otherwise perhaps you can give us the specifics of your python installation and we will try and replicate the issue locally.

On the files that under version control, we see the issue here, but the problem is that only a part of the last three files are edited automatically during setup, to remove them from version control would mean no version control on the other parts of these files, which may well be edited during development…… We will have to think what is best.
