### 2018-01-10 jronayne/PyTransport#3 @joezuntz
Thanks - the fix you suggested worked for me. Might be worth making this happen automaticaly?  The error was coming from this line in PyTransSetup.py:

    subprocess.call(["python", filename1, "install", "--home=" + location],cwd=location)

Things now continue to this error:

```
Joes-MacBook-Pro:temo jaz$ python ../test1.py.txt
running install
running build
running build_ext
building 'PyTransjoe' extension
creating build
creating build/temp.macosx-10.12-x86_64-2.7
creating build/temp.macosx-10.12-x86_64-2.7/Users
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport/PyTransport
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport/PyTransport/PyTrans
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport/PyTransport/CppTrans
creating build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport/PyTransport/CppTrans/stepper
clang -fno-strict-aliasing -fno-common -dynamic -g -O2 -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -I/usr/local/include -I/usr/local/opt/openssl/include -I/usr/local/opt/sqlite/include -I/usr/local/Cellar/python/2.7.14_2/Frameworks/Python.framework/Versions/2.7/include/python2.7 -c /Users/jaz/src/PyTransport/PyTransport/PyTrans/PyTrans.cpp -o build/temp.macosx-10.12-x86_64-2.7/Users/jaz/src/PyTransport/PyTransport/PyTrans/PyTrans.o
/Users/jaz/src/PyTransport/PyTransport/PyTrans/PyTrans.cpp:20:10: fatal error: 'numpy/arrayobject.h' file not found
#include "numpy/arrayobject.h"
         ^~~~~~~~~~~~~~~~~~~~~
1 error generated.
error: command 'clang' failed with exit status 1
```
The missing header seems to be in a fairly standard location: `/usr/local/lib/python2.7/site-packages/numpy/core/include/numpy/arrayobject.h` on my system.  I fixed this with a stackoverflow idea to do this first: `export CFLAGS="-I /usr/local/lib/python2.7/site-packages/numpy/core/include"`
It might be possible to do that automatically somehow?
