+++ AutoTest/ReadMe.md
@@ -1,0 +1,11 @@
+## Automatic Test of PyTransport 2.0
+
+Test of PyTransport distibution to ensure correct installation. This test uses the python package nose.
+Before running insure you have the correct location to the PyTransport distibution in TestSetup.py and Testrun.py.
+To excute the test run:
+
+python TestSetup.py
+
+In order to setup the test. Next run:
+
+nosetests Testrun.py
