### 2018-02-06 22e33e9
Deleted redundant tolerance module
