+++ README.md
@@ -13,7 +13,7 @@ In addition error feedback has been improved and storage of model quantities has
 PyTransport has been developed on OS X and Linux Ubuntu using Python 2.7, and is intended for use on Unix based systems.
 
 # Contibutions, Pull requests and Issues
-Any third party wishing to contribute to the PyTransport project is welcome to do so. We will attempt to implement submitted pull requests and fix issues submitted to this repository. If assistance or support is needed you can also contact us by [email](j.ronayne@qmul.ac.uk).
+Any third party wishing to contribute to the PyTransport project is welcome to do so. We will attempt to implement submitted pull requests and fix issues submitted to this repository. If assistance or support is needed you can also contact us by email at j.ronayne@qmul.ac.uk  and d.mulryne@qmul.ac.uk.
 
 # Licencing #
 PyTransport is distributed under the GNU General Public License version 3, or (at your option) any later version. This license is bundled with the source code as LICENSE.txt.
