### 2018-02-06 jronayne/PyTransport#5 @jronayne
I have now update the README.md file in the main repo to contain a section on contributions and issues.
