### 2018-02-06 jronayne/PyTransport#3 @jronayne
Glad to hear that has worked. We have added the setup.cfg file so this should no longer be an issue in the future. As you have suggested I have also added `export CFLAGS="-I /usr/local/lib/python2.7/site-packages/numpy/core/include"` to the PyTransSetup.py script so this is automated from now on. Thanks again for taking the time to find this fix and review the code.
