### 2018-02-06 #494 @jronayne
@afron @joezuntz - HI, apologies for the delay in my response. I have hopefully addressed all the feedback above so that you may continue the review.
