### 2018-03-07 #494 @joezuntz
@jronayne - many apologies for the long delay before getting back onto this.

I think this is nearly ready, just a few more issues that I'll open on the main repo.
