### 2018-03-12 3384705
Update README.md
Added quick installation guide.
