### 2018-03-12 9cee036
Update PyTransSetup.py
Fixed tabs and spacing issues.
