### 2018-03-15 #494 @joezuntz
Although some of the code methodology (such as its installation process) is non-standard for general software, the differences are within the norm for scientific software designed to support further code development by users.  Everything else is cleanly ticked off, so I'm happy to accept now.
