### 2018-03-28 #494 @jronayne
@arfon - Archive has been [created](https://zenodo.org/record/1209239#.WrufqrQdLzI) with the reviesed sofware on Zenodo.
DOI:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1209239.svg)](https://doi.org/10.5281/zenodo.1209239)

Sorry for the delay and thanks again.
