### 2018-03-28 #494 @arfon
@joezuntz - many thanks for your review here ✨

@jronayne - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00494 ⚡️:rocket: :boom:x
